pyproject.toml
[project]
name = "pdd-m5-steven-patino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.11",
]
